Completa una Serie de Datos

La función permite completar una serie de datos temporales definiendo alguna resolución disponible. Es compatible con datos agrupados por dplyr::group_by().

Uso

completar_serie(datos, fecha, resolucion, rango = range(fecha))

Argumentos

datos

tabla (data.frame, data.table, tibble) a completar.

fecha

variable de fechas.

resolucion

texto que define resolución de salida de los datos. Puede ser cualquier valor aceptado por el argumento by de la función seq.Date() o su traducción al español. Es decir, "día" (o "dia"), "semana", "mes", "trimestre" o "año", así como los plurales.

rango

un vector cuyo rango define el período a completar. Es útil si se quiere que múltiples grupos de datos tengan el mismo rango de fechas.

Valor

Devuelve un data.frame con las mismas variables de origen. La variable asociada a las fechas ahora se encuentra completa para la resolución indicada y el resto de las variables se completan con NA.

Ejemplos

# Datos de prueba completos
datos <- data.frame(fechas = seq(as.Date("1985-01-01"), as.Date("2015-12-01"),
                             by = "1 month"),
                    pp = 1)

set.seed(42)
datos_perdidos <- sample(nrow(datos), nrow(datos)/10)
datos$pp[datos_perdidos] <- NA
datos_incompletos <- na.omit(datos) #Serie de datos a completar

completar_serie(datos_incompletos, fechas, resolucion = "1 mes")
#>         fechas pp
#> 1   1985-01-01  1
#> 2   1985-02-01  1
#> 3   1985-03-01  1
#> 4   1985-04-01  1
#> 5   1985-05-01 NA
#> 6   1985-06-01  1
#> 7   1985-07-01  1
#> 8   1985-08-01  1
#> 9   1985-09-01  1
#> 10  1985-10-01  1
#> 11  1985-11-01  1
#> 12  1985-12-01  1
#> 13  1986-01-01  1
#> 14  1986-02-01  1
#> 15  1986-03-01  1
#> 16  1986-04-01  1
#> 17  1986-05-01  1
#> 18  1986-06-01  1
#> 19  1986-07-01  1
#> 20  1986-08-01 NA
#> 21  1986-09-01  1
#> 22  1986-10-01  1
#> 23  1986-11-01  1
#> 24  1986-12-01 NA
#> 25  1987-01-01  1
#> 26  1987-02-01  1
#> 27  1987-03-01  1
#> 28  1987-04-01  1
#> 29  1987-05-01  1
#> 30  1987-06-01  1
#> 31  1987-07-01  1
#> 32  1987-08-01  1
#> 33  1987-09-01  1
#> 34  1987-10-01  1
#> 35  1987-11-01  1
#> 36  1987-12-01  1
#> 37  1988-01-01  1
#> 38  1988-02-01  1
#> 39  1988-03-01  1
#> 40  1988-04-01  1
#> 41  1988-05-01  1
#> 42  1988-06-01  1
#> 43  1988-07-01  1
#> 44  1988-08-01  1
#> 45  1988-09-01  1
#> 46  1988-10-01  1
#> 47  1988-11-01  1
#> 48  1988-12-01  1
#> 49  1989-01-01 NA
#> 50  1989-02-01  1
#> 51  1989-03-01  1
#> 52  1989-04-01  1
#> 53  1989-05-01  1
#> 54  1989-06-01  1
#> 55  1989-07-01  1
#> 56  1989-08-01  1
#> 57  1989-09-01  1
#> 58  1989-10-01  1
#> 59  1989-11-01  1
#> 60  1989-12-01  1
#> 61  1990-01-01  1
#> 62  1990-02-01  1
#> 63  1990-03-01  1
#> 64  1990-04-01  1
#> 65  1990-05-01  1
#> 66  1990-06-01  1
#> 67  1990-07-01  1
#> 68  1990-08-01  1
#> 69  1990-09-01  1
#> 70  1990-10-01  1
#> 71  1990-11-01  1
#> 72  1990-12-01  1
#> 73  1991-01-01  1
#> 74  1991-02-01 NA
#> 75  1991-03-01  1
#> 76  1991-04-01  1
#> 77  1991-05-01  1
#> 78  1991-06-01  1
#> 79  1991-07-01  1
#> 80  1991-08-01  1
#> 81  1991-09-01  1
#> 82  1991-10-01  1
#> 83  1991-11-01  1
#> 84  1991-12-01  1
#> 85  1992-01-01  1
#> 86  1992-02-01  1
#> 87  1992-03-01  1
#> 88  1992-04-01  1
#> 89  1992-05-01 NA
#> 90  1992-06-01  1
#> 91  1992-07-01  1
#> 92  1992-08-01  1
#> 93  1992-09-01  1
#> 94  1992-10-01  1
#> 95  1992-11-01  1
#> 96  1992-12-01  1
#> 97  1993-01-01  1
#> 98  1993-02-01  1
#> 99  1993-03-01  1
#> 100 1993-04-01  1
#> 101 1993-05-01  1
#> 102 1993-06-01  1
#> 103 1993-07-01  1
#> 104 1993-08-01  1
#> 105 1993-09-01  1
#> 106 1993-10-01  1
#> 107 1993-11-01  1
#> 108 1993-12-01  1
#> 109 1994-01-01 NA
#> 110 1994-02-01 NA
#> 111 1994-03-01  1
#> 112 1994-04-01  1
#> 113 1994-05-01  1
#> 114 1994-06-01  1
#> 115 1994-07-01  1
#> 116 1994-08-01  1
#> 117 1994-09-01  1
#> 118 1994-10-01  1
#> 119 1994-11-01  1
#> 120 1994-12-01  1
#> 121 1995-01-01  1
#> 122 1995-02-01 NA
#> 123 1995-03-01  1
#> 124 1995-04-01  1
#> 125 1995-05-01  1
#> 126 1995-06-01  1
#> 127 1995-07-01  1
#> 128 1995-08-01 NA
#> 129 1995-09-01  1
#> 130 1995-10-01  1
#> 131 1995-11-01  1
#> 132 1995-12-01  1
#> 133 1996-01-01  1
#> 134 1996-02-01  1
#> 135 1996-03-01  1
#> 136 1996-04-01 NA
#> 137 1996-05-01  1
#> 138 1996-06-01  1
#> 139 1996-07-01  1
#> 140 1996-08-01  1
#> 141 1996-09-01  1
#> 142 1996-10-01  1
#> 143 1996-11-01  1
#> 144 1996-12-01  1
#> 145 1997-01-01  1
#> 146 1997-02-01 NA
#> 147 1997-03-01  1
#> 148 1997-04-01  1
#> 149 1997-05-01  1
#> 150 1997-06-01  1
#> 151 1997-07-01  1
#> 152 1997-08-01  1
#> 153 1997-09-01 NA
#> 154 1997-10-01  1
#> 155 1997-11-01  1
#> 156 1997-12-01  1
#> 157 1998-01-01  1
#> 158 1998-02-01 NA
#> 159 1998-03-01  1
#> 160 1998-04-01  1
#> 161 1998-05-01  1
#> 162 1998-06-01  1
#> 163 1998-07-01  1
#> 164 1998-08-01  1
#> 165 1998-09-01 NA
#> 166 1998-10-01  1
#> 167 1998-11-01  1
#> 168 1998-12-01  1
#> 169 1999-01-01  1
#> 170 1999-02-01  1
#> 171 1999-03-01  1
#> 172 1999-04-01  1
#> 173 1999-05-01  1
#> 174 1999-06-01  1
#> 175 1999-07-01  1
#> 176 1999-08-01  1
#> 177 1999-09-01  1
#> 178 1999-10-01  1
#> 179 1999-11-01  1
#> 180 1999-12-01  1
#> 181 2000-01-01  1
#> 182 2000-02-01  1
#> 183 2000-03-01  1
#> 184 2000-04-01  1
#> 185 2000-05-01  1
#> 186 2000-06-01  1
#> 187 2000-07-01  1
#> 188 2000-08-01  1
#> 189 2000-09-01  1
#> 190 2000-10-01  1
#> 191 2000-11-01  1
#> 192 2000-12-01  1
#> 193 2001-01-01  1
#> 194 2001-02-01  1
#> 195 2001-03-01  1
#> 196 2001-04-01  1
#> 197 2001-05-01 NA
#> 198 2001-06-01  1
#> 199 2001-07-01  1
#> 200 2001-08-01  1
#> 201 2001-09-01  1
#> 202 2001-10-01  1
#> 203 2001-11-01  1
#> 204 2001-12-01  1
#> 205 2002-01-01  1
#> 206 2002-02-01  1
#> 207 2002-03-01  1
#> 208 2002-04-01  1
#> 209 2002-05-01  1
#> 210 2002-06-01  1
#> 211 2002-07-01  1
#> 212 2002-08-01 NA
#> 213 2002-09-01  1
#> 214 2002-10-01  1
#> 215 2002-11-01  1
#> 216 2002-12-01  1
#> 217 2003-01-01  1
#> 218 2003-02-01  1
#> 219 2003-03-01  1
#> 220 2003-04-01  1
#> 221 2003-05-01  1
#> 222 2003-06-01  1
#> 223 2003-07-01  1
#> 224 2003-08-01  1
#> 225 2003-09-01  1
#> 226 2003-10-01  1
#> 227 2003-11-01  1
#> 228 2003-12-01 NA
#> 229 2004-01-01  1
#> 230 2004-02-01  1
#> 231 2004-03-01  1
#> 232 2004-04-01  1
#> 233 2004-05-01  1
#> 234 2004-06-01  1
#> 235 2004-07-01  1
#> 236 2004-08-01  1
#> 237 2004-09-01  1
#> 238 2004-10-01  1
#> 239 2004-11-01  1
#> 240 2004-12-01  1
#> 241 2005-01-01  1
#> 242 2005-02-01  1
#> 243 2005-03-01  1
#> 244 2005-04-01  1
#> 245 2005-05-01  1
#> 246 2005-06-01  1
#> 247 2005-07-01  1
#> 248 2005-08-01  1
#> 249 2005-09-01  1
#> 250 2005-10-01  1
#> 251 2005-11-01  1
#> 252 2005-12-01  1
#> 253 2006-01-01  1
#> 254 2006-02-01  1
#> 255 2006-03-01  1
#> 256 2006-04-01  1
#> 257 2006-05-01  1
#> 258 2006-06-01  1
#> 259 2006-07-01 NA
#> 260 2006-08-01  1
#> 261 2006-09-01  1
#> 262 2006-10-01  1
#> 263 2006-11-01  1
#> 264 2006-12-01  1
#> 265 2007-01-01  1
#> 266 2007-02-01  1
#> 267 2007-03-01  1
#> 268 2007-04-01  1
#> 269 2007-05-01  1
#> 270 2007-06-01  1
#> 271 2007-07-01  1
#> 272 2007-08-01  1
#> 273 2007-09-01  1
#> 274 2007-10-01  1
#> 275 2007-11-01  1
#> 276 2007-12-01  1
#> 277 2008-01-01  1
#> 278 2008-02-01  1
#> 279 2008-03-01  1
#> 280 2008-04-01  1
#> 281 2008-05-01  1
#> 282 2008-06-01  1
#> 283 2008-07-01 NA
#> 284 2008-08-01  1
#> 285 2008-09-01  1
#> 286 2008-10-01  1
#> 287 2008-11-01  1
#> 288 2008-12-01  1
#> 289 2009-01-01  1
#> 290 2009-02-01  1
#> 291 2009-03-01  1
#> 292 2009-04-01 NA
#> 293 2009-05-01  1
#> 294 2009-06-01  1
#> 295 2009-07-01  1
#> 296 2009-08-01  1
#> 297 2009-09-01 NA
#> 298 2009-10-01 NA
#> 299 2009-11-01 NA
#> 300 2009-12-01  1
#> 301 2010-01-01  1
#> 302 2010-02-01  1
#> 303 2010-03-01 NA
#> 304 2010-04-01  1
#> 305 2010-05-01  1
#> 306 2010-06-01  1
#> 307 2010-07-01  1
#> 308 2010-08-01  1
#> 309 2010-09-01  1
#> 310 2010-10-01  1
#> 311 2010-11-01  1
#> 312 2010-12-01  1
#> 313 2011-01-01  1
#> 314 2011-02-01 NA
#> 315 2011-03-01  1
#> 316 2011-04-01  1
#> 317 2011-05-01  1
#> 318 2011-06-01  1
#> 319 2011-07-01  1
#> 320 2011-08-01  1
#> 321 2011-09-01 NA
#> 322 2011-10-01  1
#> 323 2011-11-01  1
#> 324 2011-12-01 NA
#> 325 2012-01-01  1
#> 326 2012-02-01  1
#> 327 2012-03-01 NA
#> 328 2012-04-01  1
#> 329 2012-05-01  1
#> 330 2012-06-01  1
#> 331 2012-07-01  1
#> 332 2012-08-01  1
#> 333 2012-09-01  1
#> 334 2012-10-01  1
#> 335 2012-11-01  1
#> 336 2012-12-01  1
#> 337 2013-01-01  1
#> 338 2013-02-01  1
#> 339 2013-03-01  1
#> 340 2013-04-01  1
#> 341 2013-05-01  1
#> 342 2013-06-01  1
#> 343 2013-07-01  1
#> 344 2013-08-01  1
#> 345 2013-09-01  1
#> 346 2013-10-01  1
#> 347 2013-11-01 NA
#> 348 2013-12-01 NA
#> 349 2014-01-01  1
#> 350 2014-02-01  1
#> 351 2014-03-01  1
#> 352 2014-04-01 NA
#> 353 2014-05-01  1
#> 354 2014-06-01  1
#> 355 2014-07-01  1
#> 356 2014-08-01 NA
#> 357 2014-09-01  1
#> 358 2014-10-01  1
#> 359 2014-11-01 NA
#> 360 2014-12-01  1
#> 361 2015-01-01  1
#> 362 2015-02-01 NA
#> 363 2015-03-01  1
#> 364 2015-04-01  1
#> 365 2015-05-01  1
#> 366 2015-06-01  1
#> 367 2015-07-01 NA
#> 368 2015-08-01  1
#> 369 2015-09-01  1
#> 370 2015-10-01  1
#> 371 2015-11-01  1