Import Experimental Data from agroclimR Data Workbook

import_exp_data() performs extraction of experimental data from the agroclimR data workbook, including location, soil, climate data, and observed data for phenology, leaf area index, dry matter, and yield.

Usage

import_exp_data(files, model = "oryza")

Arguments

files

Vector of strings with workbook names (.xls extension file included).

model

String indicating the model name, default "oryza",options: ("dssat", "aquacrop").

Value

List containing raw and observed data by component ("data", "soil", "wth", "phen", "lai", "dry_matter", "yield").

See also

Examples

#' # File names vector, extension include
name_file = c("SDTO_F2000.xlsx")

# Files directory
test_file = system.file("extdata", name_file[1], package = "agroclimR")

# Import data to R lists and tibble formats
obs_data = import_exp_data(test_file, model = "oryza")

head(obs_data)
#> $data
#> # A tibble: 1 × 2
#>   file                                                              input_data  
#>   <chr>                                                             <list>      
#> 1 /home/runner/work/_temp/Library/agroclimR/extdata/SDTO_F2000.xlsx <named list>
#> 
#> $soil
#> # A tibble: 1 × 5
#>   site  soil                lat   lon  elev
#>   <chr> <list>            <dbl> <dbl> <dbl>
#> 1 SDTO  <tibble [3 × 20]>  3.91 -75.0   415
#> 
#> $wth
#> # A tibble: 1 × 6
#>   site  ws_id wth                    lat   lon  elev
#>   <chr> <dbl> <list>               <dbl> <dbl> <dbl>
#> 1 SDTO      1 <tibble [1,461 × 7]>  3.91 -75.0   415
#> 
#> $phen
#> # A tibble: 21 × 4
#>    exp_file               data             var   value
#>    <chr>                  <list>           <chr> <dbl>
#>  1 SDTO_FED2000_MADRI_S1  <tibble [5 × 2]> IDAT     51
#>  2 SDTO_FED2000_MADRI_S1  <tibble [5 × 2]> FDAT     81
#>  3 SDTO_FED2000_MADRI_S1  <tibble [5 × 2]> MDAT    104
#>  4 SDTO_FED2000_MADRII_S1 <tibble [5 × 2]> IDAT     48
#>  5 SDTO_FED2000_MADRII_S1 <tibble [5 × 2]> FDAT     85
#>  6 SDTO_FED2000_MADRII_S1 <tibble [5 × 2]> MDAT    115
#>  7 SDTO_FED2000_MADRI_S2  <tibble [5 × 2]> IDAT     49
#>  8 SDTO_FED2000_MADRI_S2  <tibble [5 × 2]> FDAT     85
#>  9 SDTO_FED2000_MADRI_S2  <tibble [5 × 2]> MDAT    105
#> 10 SDTO_FED2000_MADRII_S2 <tibble [5 × 2]> IDAT     40
#> # ℹ 11 more rows
#> 
#> $lai
#> # A tibble: 77 × 5
#>    exp_file              date       var   value    se
#>    <chr>                 <date>     <chr> <dbl> <dbl>
#>  1 SDTO_FED2000_MADRI_S1 2013-05-21 LAI      NA    NA
#>  2 SDTO_FED2000_MADRI_S1 2013-05-28 LAI      NA    NA
#>  3 SDTO_FED2000_MADRI_S1 2013-06-07 LAI      NA    NA
#>  4 SDTO_FED2000_MADRI_S1 2013-06-12 LAI      NA    NA
#>  5 SDTO_FED2000_MADRI_S1 2013-06-18 LAI      NA    NA
#>  6 SDTO_FED2000_MADRI_S1 2013-06-25 LAI      NA    NA
#>  7 SDTO_FED2000_MADRI_S1 2013-07-07 LAI      NA    NA
#>  8 SDTO_FED2000_MADRI_S1 2013-07-10 LAI      NA    NA
#>  9 SDTO_FED2000_MADRI_S1 2013-07-17 LAI      NA    NA
#> 10 SDTO_FED2000_MADRI_S1 2013-07-22 LAI      NA    NA
#> # ℹ 67 more rows
#> 
#> $dry_matter
#> # A tibble: 385 × 5
#>    exp_file              date       var   value    se
#>    <chr>                 <date>     <chr> <dbl> <dbl>
#>  1 SDTO_FED2000_MADRI_S1 2013-05-21 WAGT   610.  37.2
#>  2 SDTO_FED2000_MADRI_S1 2013-05-21 WLVD     0    0  
#>  3 SDTO_FED2000_MADRI_S1 2013-05-21 WLVG   316.  20.4
#>  4 SDTO_FED2000_MADRI_S1 2013-05-21 WSO      0    0  
#>  5 SDTO_FED2000_MADRI_S1 2013-05-21 WST    294.  20  
#>  6 SDTO_FED2000_MADRI_S1 2013-05-28 WAGT  1188  254  
#>  7 SDTO_FED2000_MADRI_S1 2013-05-28 WLVD     0    0  
#>  8 SDTO_FED2000_MADRI_S1 2013-05-28 WLVG   544. 122. 
#>  9 SDTO_FED2000_MADRI_S1 2013-05-28 WSO      0    0  
#> 10 SDTO_FED2000_MADRI_S1 2013-05-28 WST    644. 132. 
#> # ℹ 375 more rows
#>